Change source and target to Java versions to 1.7
The plugin compiles fine with JDK 1.7 and does not seem to use any 1.8-only syntax/functionailty, as far as I can see. It runs just fine when compiled with JDK 1.7 as well. Maven failed when building this plugin on Ubuntu 14.04, as only OpenJDK 1.7 was installed; 1.8 is not yet available in the Ubuntu package repositories. Updating pom.xml to allow for compilation with JDK 1.7.